Clara: TextFlow: fixed issue with word wrapping
The issue with word wrapping was that in case that m_pos + width was
greater or equal to m_end it fallbacked to m_end-m_pos branch which
caused that printed string was greater than configured width.

Also caused duplication of words as in next iteration m_len was
added to m_pos this was less then m_end.

Originally reported in https://github.com/catchorg/Catch2/issues/1400